Managing SSL certificates & private keys in Puppet
Real private key will be managed with hiera-eyaml, but private key used
with bogus self-signed certificate for vagrant need not be (and cannot
be) secured.

Just in case I'm doing something really stupid, the current
'profile::ssl-cert::secret-key-wiki-jira' entry is a bogus private key,
which is nonetheless safe to merge because edamame currently doesn't
serve https.

If this use of multiline string in hiera-eyaml is validated, I'll
replace the entry with the real private key.