Remove hacky hard-coding of IP addresses when binding confluence and confluence-cache together
This took me an awful long time to figure out, the IP addresses used on the
production host didn't really guarantee that they would be there after reboots,
etc.

I had previously misattributed an availability failure of the recent confluence
container tag to be the container, but really it was this hard-coded IP address
which was no longer correct.

By using --link we ensure that the confluence-cache container gets updated with
the appropriate /etc/hosts information to hit confluence:8080