Introduce a Mac buildnode role.
This is the first role that doesn't have a vegetable associated with it. As we
move towards more of a cloud-based infrastructure, our roles need to slowly
morph into the roles for a node instead of our "pet" machines we have right now.

This commit makes provisioning Docker on a profile::buildslave optional, since
Mac OS X is not going to run docker.

[FIX INFRA-601]