Haphazard attempt to get jenkins::credentials to work
I'm committing this for posterity, but plan on reverting it with the next
commit.

Basically Jenkins's security cannot be modeled in an idempotent fashion. Since
we cannot pre-share any public keys with it, we have to go through this process:

 * Stand up Jenkins unsecured (but obviously walled off)
 * Set up authentication, authorization, leaving CLI access for anonymous
 * Create a role account and give it our SSH public key
 * Remove CLI access for anonymous and grant it for our role account

This is a lot of gnarly work and needs to be made better in Jenkins itself. A
temporary workaround is to allow CLI access for anonymous, but only from the
lo0 interface on the system