Upgrade the garethr/docker module to what is "upstream" and set engine version to 1.12
Tested this manually with Vagrant/serverspec by provisioning a machine with the
current production branch and then re-running Puppet with this change

The only potential issue seen was that switching from devicemapper to aufs
requires the daemon to be restarted before any pulls can happen. This might make
sense to do as a manual step as needed.