Speeding up rsync run
Make rsync preserve timestamp during copy, so that successive runs can
speed up. In the likely event that files haven't been touched, this will
allow rsync to skip computing checksums.

This part of rsync takes forever trying to compute checksums over large
number of files, slowing the release tremendously.

Tyler has some concerns that this might not be as effective, citing that
the upstream UC generation that places plugins might be touching file
stamps unnecessarily, but looking at actual files in these directories,
they are not. And even if that is the case, this addition will merely
make no improvements and will not make situation any worse.