Add a Kubernetes parameter to define where a deploy is happening.
This gives us the option to alter how resources are deployed depending
on the environment in which it occurs.  Currently, certain containers
will fail to boot, as they expect to use Azure storage volumes.  With
this change, we can optionally switch containers to be deployed with
ephemeral storage, enabling development with a local Kubernetes cluster.