Redirect governance documents from wiki to jenkins.io
The governance document and governance meeting pages have already migrated
and are only pointers to the official jenkins.io pages.

The governance board page has been migrated to jenkins.io but does not
have a pointer in the page.  This redirect avoids editing that page by
redirecting to jenkins.io.

The candidacy process and election process for the Jenkins board are
maintained on jenkins.io.  No need to show the user an outdated copy
from the wiki.