Replace markewaite ED25519 public key with RSA key
No indicator on any of the keys for the type of key.  All the other keys
appear to be RSA keys.  Likely that the key is being listed as an RSA
key even when it is an ED25519 key.  Should have followed the standard
pattern in the file and used an RSA key originally.