Add confluence URL checks in CI
Keep the checks lightweight by recording the expected URLs in local files,
one for the URLs that are known to exist in Confluence and one for the
URLs that are known to not exist in Confluence.

Skips security URLs and any URL that includes a literal '(' because there
are relatively few of them and they are painful to handle in the simple
checks this is performing.