Fix JenkinsRule test failures - install fontconfig pkg
The JDK reports a null pointer exception during some JenkinsRule
tests if the fontconfig package is not installed on Linux computers.
The fontconfig package is installed in our Docker controller and agent
images.  https://github.com/jenkins-infra/packer-images/pull/86
added it to the packer image definition files.  The
https://ci.jenkins.io/job/Infra/job/packer-images/job/main/189/
job built that image using
https://github.com/jenkins-infra/packer-images/commit/421f74a2a889242fd8934c6589c9d91c42f632c1

This pull request extracted the AMI ID from that job and inserted it
for the amd64 and arm64 images.